Do not spawn client-side particles with the server's world
Usage of the proxy classes to spawn particles currently allows
the server's world to be easily leaked into the client side. It is
not valid to use any server objects on a side other than the
server.

This patch removes the World argument from the methods relevant to
spawning particles in Ice and Fire's proxies and instead grabs
it from `Minecraft#getInstance()`.

Fixes #1558, #1500, #1543, and #1530.