Turn Gorgon Temple into a Jigsaw Structure
Resolves: https://github.com/Alex-the-666/Ice_and_Fire/issues/3678

Basically that issue was caused by the terraforming that Minecraft does being applied to every nbt piece and causing any air space it makes below sealevel to be filled with water. Which then triggers a vanilla bug where it waterlogs the structure's own pieces as a result. Not good.

By turning the temple into a Jigsaw structure, now the terraforming only happens at the base of the exposed temple which keeps the basement dry.

Things to note:
The temple is split up into 3 nbt files. The building, the basement, and the gorgon. Each has their own pool. The building has a downward jigsaw block that connects to the basement's upward jigsaw block. I had to put 2 piston/redstone block under two of the always on switches due to redstone dust not properly updating its state when crossing between jigsaw pieces. The piston seems to keep its state just fine.

The basement has a second upward jigsaw block that points to the gorgon pool that will spawn the gorgon nbt file with the mob (that has persistence set to true). I did a pool for the mob so you can add more entries to the pool to randomize what mob spawns or what kind of gorgon. Could even randomize what status buff the gorgon has.

The gorgon piece class is deleted as it is now useless. Instead, I made a dummy piece class to keep under the old gordon piece class registry name. This is to prevent logspam for existing worlds. However, you can remove this dummy piece and put in a log filter to silence the logspam if it is your piece like this: https://github.com/TelepathicGrunt/RepurposedStructures/blob/6b730683e6ce950846e853fb2d25a41e7b77d2e2/src/main/java/com/telepathicgrunt/repurposedstructures/utils/LogSpamFiltering.java#L56

Otherwise, this should fix that waterlogging bug while also giving you now the opportunity to randomize the temple pieces. Heck, you can now add on processors easier and make it randomize the blocks in the temple so it isn't the same every single time.

Hope this helps! be sure to test this first. I couldn't run Ice and Fire in dev due to missing client classes (TabulaModelAccessor) from the github. So I copied the temple files to my other test mod and made sure it was working there. But again, double check my work in the actual ice and fire mod but it should work just fine.