feat: Optimise WorldUtil (#5513)
* Optimise WorldUtil#canGenerate

When `checkFluid` is true and there is a fluid at the origin, it returns false. We can skip a bunch of redundant calculations in this case by reordering the code.

* Optimise WorldUtil#isEntityChunkLoaded

Skips a redundant `instanceof` check when the chunk isn't loaded

* Optimise WorldUtil#isChunkLoaded

Avoids calling `world.getChunkSource()` twice by using Java instanceof pattern matching.